requirements.txt
groq>=0.14.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
python-docx>=1.1.0
PyPDF2>=3.0.0
python-magic-bin>=0.4.14
lxml>=5.0.0
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.0
bcrypt>=4.1.2
PyJWT>=2.8.0
Pillow>=10.1.0
